+++ spec/Lungo.Core.Spec.js
@@ -1,9 +1,0 @@
-describe('Lungo Core', function() {
-    it('', function() {
-
-    });
-    
-    it('', function() {
-        
-    });
-});