+++ src/boot/Lungo.Boot.Article.js
@@ -10,8 +10,10 @@
 
 LUNGO.Boot.Article = (function(lng, undefined) {
 
+    var ATTRIBUTE = lng.Constants.ATTRIBUTE;
+    var ELEMENT = lng.Constants.ELEMENT;
     var SELECTORS = {
-        LIST_IN_ARTICLE: 'article.list',
+        LIST_IN_ARTICLE: 'article.list, aside.list',
         SCROLL_IN_ARTICLE: '.scrollable',
         CHECKBOX_IN_ARTICLE: '.checkbox, .radio'
     };
@@ -38,18 +40,18 @@ LUNGO.Boot.Article = (function(lng, undefined) {
 
     var _createListElement = function(article) {
         if (article.children().length === 0) {
-            var article_id = article.attr('id');
-            article.append('<ul id="' + article_id + '_list"></ul>');
+            var article_id = article.attr(ATTRIBUTE.ID);
+            article.append(ELEMENT.LIST);
         }
     };
 
     var _createScrollElement = function(scroll) {
-        var scroll_id = scroll.attr('id');
-        lng.View.Scroll.create(scroll_id);
+        var scroll_id = scroll.attr(ATTRIBUTE.ID);
+        lng.View.Scroll.init(scroll_id);
     };
 
     var _createCheckboxElement = function(checkbox) {
-        checkbox.append('<span>&nbsp;</span>');
+        checkbox.append(ELEMENT.SPAN);
     };
 
     return {