+++ src/view/Lungo.View.Template.Binding.js
@@ -10,36 +10,27 @@
 
 LUNGO.View.Template.Binding = (function(lng, undefined) {
 
-    var BINDING_START = '{{';
-    var BINDING_END = '}}';
-    var BINDING_PARSER = /\{{.*?\}}/gi;
+    var BINDING = lng.Constants.BINDING;
+    var ERROR = lng.Constants.ERROR;
 
     /**
      * Performs databinding process for a data set and a given template
      *
      * @method create
      *
-     * @param {String} Id of the container showing the result of databinding
      * @param {String} Databinding Template Id
      * @param {Object} Data for binding
-     * @param {Function} Callback when the process is complete
      */
-    var create = function(container_id, template_id, data, callback) {
-        if (lng.View.Template.exists(template_id)) {
-            var template = lng.View.Template.get(template_id);
-            var markup = _processData(data, template);
-            _render(container_id, markup);
-            lng.Core.execute(callback);
-        } else {
-            lng.Core.log(3, 'lng.View.Template.binding: id ' + template_id + ' not exists');
-        }
+    var create = function(template_id, data) {
+        var template = lng.View.Template.get(template_id);
+        return _processData(data, template);
     };
 
     var dataAttribute = function(element, attribute) {
         var data = element.data(attribute.tag);
 
         if (data) {
-            var html_binded = attribute.html.replace(BINDING_START + 'value' + BINDING_END, data);
+            var html_binded = attribute.html.replace(BINDING.START + BINDING.KEY + BINDING.END, data);
             element.prepend(html_binded);
         }
     };
@@ -52,7 +43,7 @@ LUNGO.View.Template.Binding = (function(lng, undefined) {
         } else if (data_type === 'object') {
             return _bindProperties(data, template);
         } else {
-            lng.Core.log(3, 'View.Template ERROR >> No type defined.');
+            lng.Core.log(3, ERROR.BINDING_DATA_TYPE);
         }
     };
 
@@ -67,8 +58,8 @@ LUNGO.View.Template.Binding = (function(lng, undefined) {
     var _bindProperties = function(element, template) {
         var binding_field;
         for (var property in element) {
-            if (lng.Core.isOwnProperty(element, property)) {
-                binding_field = new RegExp(BINDING_START + property + BINDING_END, 'g');
+            if (lng.Core.isOwnProperty(element, property) && element[property] !== null) {
+                binding_field = new RegExp(BINDING.START + property + BINDING.END, 'g');
                 template = template.replace(binding_field, element[property]);
             }
         }
@@ -76,12 +67,7 @@ LUNGO.View.Template.Binding = (function(lng, undefined) {
     };
 
     var _removeNoBindedProperties = function(template) {
-        return template.replace(BINDING_PARSER, '');
-    };
-
-    var _render = function(container_id, markup) {
-        var container = lng.dom('#' + container_id);
-        container.html(markup);
+        return template.replace(BINDING.PARSER, '');
     };
 
     return {