+++ src/view/Lungo.View.Template.js
@@ -10,6 +10,8 @@
 
 LUNGO.View.Template = (function(lng, undefined) {
 
+    var ERROR = lng.Constants.ERROR;
+
     var _templates = {};
 
     /**
@@ -51,22 +53,43 @@ LUNGO.View.Template = (function(lng, undefined) {
     /**
      * Performs databinding process for a data set and a given template
      *
-     * @method binding
+     * @method render
      *
-     * @param {String} Id of the container showing the result of databinding
+     * @param {String} Element selector for showing the result of databinding
      * @param {String} Databinding Template Id
      * @param {Object} Data for binding
      * @param {Function} Callback when the process is complete
      */
-    var binding = function(container_id, template_id, data, callback) {
-        lng.View.Template.Binding.create(container_id, template_id, data, callback);
+    var render = function(element, template_id, data, callback) {
+        if (lng.View.Template.exists(template_id)) {
+            var container = lng.dom(element);
+            var markup = this.markup(template_id, data);
+            container.html(markup);
+
+            lng.Core.execute(callback);
+        } else {
+            lng.Core.log(3, ERROR.BINDING_TEMPLATE + template_id);
+        }
+    };
+
+    /**
+     * Performs databinding process for a data set and a given template
+     *
+     * @method markup
+     *
+     * @param {String} Databinding Template Id
+     * @param {Object} Data for binding
+     */
+    var markup = function(template_id, data) {
+        return lng.View.Template.Binding.create(template_id, data);
     };
 
     return {
         create: create,
         exists: exists,
         get: get,
-        binding: binding
+        render: render,
+        markup: markup
     };
 
 })(LUNGO);